backend/requirements.txt
# Web Framework
fastapi==0.121.2
uvicorn[standard]==0.38.0

# Database
SQLAlchemy==2.0.44

# Validation & Settings
pydantic==2.12.4
pydantic-settings==2.12.0
email-validator==2.3.0

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.20

# Rate Limiting
slowapi==0.1.9

# Email
aiosmtplib==3.0.2
jinja2==3.1.5

# Redis for caching
redis==5.2.1
hiredis==3.1.0